+++ cupy_backends/cuda/api/runtime.pyx
@@ -62,7 +62,7 @@ cdef extern from *:
     ctypedef StreamCallbackDef* StreamCallback 'cudaStreamCallback_t'
 
 
-cdef extern from 'cupy_cuda_runtime.h' nogil:
+cdef extern from '../cupy_cuda_runtime.h' nogil:
 
     # Types
     ctypedef struct _PointerAttributes 'cudaPointerAttributes':
