+++ cupy_backends/cuda/cupy_rtc.h
@@ -2,7 +2,7 @@
 #define INCLUDE_GUARD_CUPY_RTC_H
 
 #if CUPY_USE_HIP
-#include "cupy_hiprtc.h"
+#include "hip/cupy_hiprtc.h"
 #else
 #include "cupy_nvrtc.h"
 #endif
