+++ cupy_backends/cuda/hip/cupy_runtime.h
@@ -1,14 +1,10 @@
-// This file is a stub header file of hip for Read the Docs.
-
-#ifndef INCLUDE_GUARD_CUPY_HIP_RUNTIME_H
-#define INCLUDE_GUARD_CUPY_HIP_RUNTIME_H
+#ifndef INCLUDE_GUARD_HIP_CUPY_RUNTIME_H
+#define INCLUDE_GUARD_HIP_CUPY_RUNTIME_H
 
 #include <hip/hip_runtime_api.h>
-#include "../cupy_hip_common.h"
+#include "cupy_common.h"
 
 extern "C" {
-
-bool hip_environment = true;
 
 // Error handling
 const char* cudaGetErrorName(cudaError_t hipError) {
@@ -80,6 +76,35 @@ cudaError_t cudaDeviceGetLimit(size_t* pValue, cudaLimit limit) {
 cudaError_t cudaDeviceSetLimit(cudaLimit limit, size_t value) {
     // see https://github.com/ROCm-Developer-Tools/HIP/issues/1632
     return hipErrorUnknown;
+}
+
+// IPC operations
+cudaError_t cudaIpcCloseMemHandle(void* devPtr) {
+    return hipIpcCloseMemHandle(devPtr);
+}
+
+cudaError_t cudaIpcGetEventHandle(cudaIpcEventHandle_t* handle, cudaEvent_t event) {
+    return hipErrorUnknown;
+
+    // TODO(leofang): this is supported after ROCm-Developer-Tools/HIP#1996 is released;
+    // as of ROCm 3.5.0 it is still not supported
+    //return hipIpcGetEventHandle(handle, event);
+}
+
+cudaError_t cudaIpcGetMemHandle(cudaIpcMemHandle_t* handle, void* devPtr) {
+    return hipIpcGetMemHandle(handle, devPtr);
+}
+
+cudaError_t cudaIpcOpenEventHandle(cudaEvent_t* event, cudaIpcEventHandle_t handle) {
+    return hipErrorUnknown;
+
+    // TODO(leofang): this is supported after ROCm-Developer-Tools/HIP#1996 is released;
+    // as of ROCm 3.5.0 it is still not supported
+    //return hipIpcOpenEventHandle(event, handle);
+}
+
+cudaError_t cudaIpcOpenMemHandle(void** devPtr, cudaIpcMemHandle_t handle, unsigned int flags) {
+    return hipIpcOpenMemHandle(devPtr, handle, flags);
 }
 
 // Memory management
@@ -315,4 +340,4 @@ cudaError_t cudaDestroySurfaceObject(cudaSurfaceObject_t surfObject) {
 
 } // extern "C"
 
-#endif // #ifndef INCLUDE_GUARD_CUPY_HIP_RUNTIME_H
+#endif // #ifndef INCLUDE_GUARD_HIP_CUPY_RUNTIME_H
