+++ cupy_backends/cuda/libs/curand.pyx
@@ -10,7 +10,7 @@ from cupy_backends.cuda cimport stream as stream_module
 # Extern
 ###############################################################################
 
-cdef extern from 'cupy_curand.h' nogil:
+cdef extern from '../cupy_curand.h' nogil:
     # Generator
     int curandCreateGenerator(Generator* generator, int rng_type)
     int curandDestroyGenerator(Generator generator)
