+++ cupy_backends/cuda/libs/cusolver.pyx
@@ -17,7 +17,7 @@ cdef extern from '../cupy_cuComplex.h':
     ctypedef struct cuDoubleComplex 'cuDoubleComplex':
         double x, y
 
-cdef extern from 'cupy_cusolver.h' nogil:
+cdef extern from '../cupy_cusolver.h' nogil:
     # Context
     int cusolverDnCreate(Handle* handle)
     int cusolverSpCreate(SpHandle* handle)
