+++ cupy_backends/cuda/libs/profiler.pyx
@@ -4,7 +4,7 @@
 from cupy_backends.cuda.api cimport runtime
 
 
-cdef extern from 'cupy_profiler.h' nogil:
+cdef extern from '../cupy_profiler.h' nogil:
     runtime.Error cudaProfilerInitialize(const char *configFile,
                                          const char *outputFile,
                                          int outputMode)
