+++ cupy_setup_build.py
@@ -90,7 +90,7 @@ cuda_files = [
 
 if use_hip:
     # We handle nvtx (and likely any other future support) here, because
-    # the HIP stubs (cupy_hip.h/cupy_hip_common.h) would cause many symbols
+    # the HIP stubs (hip/cupy_*.h) would cause many symbols
     # to leak into all these modules even if unused. It's easier for all of
     # them to link to the same set of shared libraries.
     MODULES.append({
