+++ install/build.py
@@ -162,6 +162,8 @@ def get_compiler_setting(use_hip):
         if os.path.exists(nvtoolsext_path):
             include_dirs.append(os.path.join(nvtoolsext_path, 'include'))
             library_dirs.append(os.path.join(nvtoolsext_path, 'lib', 'x64'))
+        else:
+            define_macros.append(('CUPY_NO_NVTX', '1'))
 
     # For CUB, we need the complex and CUB headers. The search precedence for
     # the latter is:
